implementing product factors and partial factor application for the new Factor space.
factor marginalisation is WIP. it  needs two separate cases.
1) marginalising over a DiscreteRandomVar
	get a list of all possible initalisations for the remaining DiscreteRandomVar within scope.
	then for/hash over this list i
	(values i (flsum (map (x) (hash-ref old-data (set-union x i)) var-labels)

	or

	(values i (map (x) (cons weight . canonical-factor))  this needs to be recurisve.

that needs a lot more thought.  now that I think about it, product factors needs to be re-done.
as do partial applications.  both of these were implemented by nulling the factor-data and
instead playing with function applications.  this won't work, as factor marginalisation must be over the data.

the commited code compiles but I've not tested it.

	modified:   factors.rkt
	modified:   random-var.rkt