changing product factors
previously when multiplying a CanonicalFactor by a float
inside of a Factor, we "merged" the float into CanonicalFactor
form.  We now keep the float outside in the CanonicalMixture
list.

This allows for stupid/naive approximations of CanonicalMixtures
where we simply drop the CanonicalFactors with the smallest
abs coefficient.

WIP marginalizing across discrete var in a Factor

	modified:   factor-tests.rkt
	modified:   factors.rkt