factor marginalisation is complete.
I had quite a bit of trouble with the type checker here.  In the past, I set FactorData to (HashTable (Set TableCPDIndex) (U Float CanonicalMixture)).  I mentioned that that was probably a mistake, and the U should probably be about the HashTable.  This indeed came back to bite me.

In marginalising discrete vars, the function would look up things in the old FactorData.  We know that everything in there is either a Float or a CanonicalMixture, but the compiler doesn't, so it always returns (U Float CanonicalMixture).  Since we are looking across the entire scope of the discrete var, we end up with a (Listof (U Float CanonicalMixture)).

Undoing that union was quite problematic.  Getting it to run the float case was easy.  Naively, I thought I could convert (Listof (U Float CanonicalMixture)) -> (Listof CanonicalMixture) just by doing (filter CanonicalMixture? _)

Alas, one cannot define that predicate, as CanonicalMixture = (Listof (Pairof Float CanonicalFactor)).  CanonicalFactor is an executable struct, and typed/racket complains when you try doing fancy stuff to that type.

Worse, (filter (curry not float?) _) would give a (Listof Any).  We would always lose type information with a filter that wasn't looking for CanonicalMixture exactly.

I tried to define my own predicate, but the type checker is not capable of combine information about several parts of the input.  A future version may work, a kind person on IRC suggests https://github.com/racket/typed-racket/pull/475 has been merged and is in the nightly build.

So after a couple hours of frustration, I defined a list called not-floats, which used for/list with a #:when (not (float? _)) guard.  This was able to keep the correct type information.

I wasn't sure if this was me being an idiot or a rough edge on typed racket.  In any case I felt it necessary to document it.  So I put it in a git commit that no one will read (including myself).

	modified:   factors.rkt